requirements.txt
# Core dependencies
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Classic ML
xgboost>=2.0.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0

# API (optional - uncomment if needed)
# fastapi>=0.100.0
# uvicorn>=0.23.0
# pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Graph processing
networkx>=3.1

# Document generation
python-docx>=0.8.11
